+++ src/components/param-selector-view/index.tsx
@@ -1,16 +1,13 @@
-import React, { ChangeEvent, useEffect, useRef, useState } from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 import { RmgCard, RmgLoader, RmgPage } from '@railmapgen/rmg-components';
 import { useSearchParams } from 'react-router-dom';
-import { Button, Container, Heading, HStack, SystemStyleObject, useOutsideClick, VStack } from '@chakra-ui/react';
+import { Container, Heading, HStack, SystemStyleObject, useOutsideClick, useToast } from '@chakra-ui/react';
 import { useTranslation } from 'react-i18next';
-import { MdAdd, MdInsertDriveFile, MdOpenInBrowser, MdUpload } from 'react-icons/md';
-import { nanoid } from 'nanoid';
 import rmgRuntime from '@railmapgen/rmg-runtime';
 import { Events, LocalStorageKey, ParamConfig } from '../../constants/constants';
 import ParamSelector from '../param-selector-view/param-selector';
-import { getParamRegistry, importParam } from '../../util/param-manager-utils';
-import TemplateModal from '../modal/template-modal';
-import { readFileAsText } from '../../util/utils';
+import { getParamRegistry } from '../../util/param-manager-utils';
+import SelectorActions from './selector-actions';
 
 const paramSelectorCardStyle: SystemStyleObject = {
     flexDirection: 'column',
@@ -37,14 +34,14 @@ const paramSelectorCardStyle: SystemStyleObject = {
 export default function ParamSelectorView() {
     const { t } = useTranslation();
 
-    const [searchParams, setSearchParams] = useSearchParams();
+    const [searchParams] = useSearchParams();
     const urlParamId = searchParams.get('project');
 
     const [paramRegistry, setParamRegistry] = useState<ParamConfig[]>([]);
     const [selectedParam, setSelectedParam] = useState<string>();
-    const [isTemplateModalOpen, setIsTemplateModalOpen] = useState(false);
     const selectorRef = useRef<HTMLDivElement>(null);
-    const fileInputRef = useRef<HTMLInputElement>(null);
+
+    const toast = useToast();
 
     useEffect(() => {
         // init paramRegistry state once
@@ -52,47 +49,6 @@ export default function ParamSelectorView() {
     }, []);
 
     useOutsideClick({ ref: selectorRef, handler: () => setSelectedParam(undefined) });
-
-    const handleNew = () => {
-        setSearchParams({ project: nanoid() });
-        rmgRuntime.event(Events.NEW_PARAM, {});
-    };
-
-    const handleOpenTemplate = (param: Record<string, any>) => {
-        const id = importParam(JSON.stringify(param));
-        setSearchParams({ project: id });
-    };
-
-    const handleImportProject = async (event: ChangeEvent<HTMLInputElement>) => {
-        const file = event.target.files?.[0];
-        console.log('handleImportProject():: received file', file);
-
-        try {
-            if (file?.type !== 'application/json') {
-                // TODO
-                // dispatch(setGlobalAlert({ status: 'error', message: t('OpenActions.invalidType') }));
-            }
-
-            const paramStr = await readFileAsText(file!);
-            const id = importParam(paramStr);
-            setSearchParams({ project: id });
-            rmgRuntime.event(Events.UPLOAD_PARAM, {});
-        } catch (err) {
-            // TODO
-            // dispatch(setGlobalAlert({ status: 'error', message: t('OpenActions.unknownError') }));
-            console.error('handleImportProject():: Unknown error occurred while parsing the uploaded file', err);
-        }
-
-        // clear field for next upload
-        event.target.value = '';
-    };
-
-    const handleOpenSelected = () => {
-        if (selectedParam) {
-            setSearchParams({ project: selectedParam });
-            rmgRuntime.event(Events.OPEN_PARAM, {});
-        }
-    };
 
     const handleDelete = (id: string) => {
         window.localStorage.removeItem(LocalStorageKey.PARAM_BY_ID + id);
@@ -104,7 +60,9 @@ export default function ParamSelectorView() {
         rmgRuntime.event(Events.REMOVE_PARAM, {});
     };
 
-    const isProjectLimitReached = paramRegistry.length >= 10;
+    const handleError = (message: string) => {
+        toast({ description: message, status: 'error', duration: 10000, isClosable: true });
+    };
 
     return (
         <RmgPage justifyContent="center">
@@ -123,50 +81,14 @@ export default function ParamSelectorView() {
                             onParamRemove={handleDelete}
                         />
 
-                        <VStack>
-                            <Button leftIcon={<MdAdd />} onClick={handleNew} isDisabled={isProjectLimitReached}>
-                                {t('Blank project')}
-                            </Button>
-                            <Button
-                                leftIcon={<MdInsertDriveFile />}
-                                onClick={() => setIsTemplateModalOpen(true)}
-                                isDisabled={isProjectLimitReached}
-                            >
-                                {t('Open template')}
-                            </Button>
-                            <Button
-                                leftIcon={<MdUpload />}
-                                onClick={() => fileInputRef.current?.click()}
-                                isDisabled={isProjectLimitReached}
-                            >
-                                {t('Import project')}
-                            </Button>
-                            <Button
-                                colorScheme="primary"
-                                leftIcon={<MdOpenInBrowser />}
-                                onClick={handleOpenSelected}
-                                isDisabled={selectedParam === undefined}
-                            >
-                                {t('Open selected')}
-                            </Button>
-                        </VStack>
+                        <SelectorActions
+                            selectedParam={selectedParam}
+                            disableNew={paramRegistry.length >= 10}
+                            onError={handleError}
+                        />
                     </HStack>
                 </RmgCard>
             </Container>
-
-            <input
-                ref={fileInputRef}
-                type="file"
-                accept=".json"
-                hidden={true}
-                onChange={handleImportProject}
-                data-testid="file-upload"
-            />
-            <TemplateModal
-                isOpen={isTemplateModalOpen}
-                onClose={() => setIsTemplateModalOpen(false)}
-                onOpenParam={handleOpenTemplate}
-            />
         </RmgPage>
     );
 }
