+++ src/i18n/translations/zh-Hans.json
@@ -301,6 +301,7 @@
     "Unable to connect to main line.": "无法连接至主线。",
     "Unable to disconnect from main line.": "无法从主线断开。",
     "Unable to remove this station.": "无法删除此车站。",
+    "Unknown": "未知",
     "View": "查看",
     "Visit": "访问",
     "Visit GitHub": "访问GitHub",
