+++ src/i18n/translations/zh-Hant.json
@@ -301,6 +301,7 @@
     "Unable to connect to main line.": "無法連接至主綫。",
     "Unable to disconnect from main line.": "無法從主綫斷開。",
     "Unable to remove this station.": "無法移除該車站。",
+    "Unknown": "未知",
     "View": "檢視",
     "Visit": "造訪",
     "Visit GitHub": "造訪GitHub",
